Move core.transformations to lib.transformations
Part of Issue #287
