+++ package/doc/sphinx/source/documentation_pages/lib_modules.rst
@@ -1,16 +1,19 @@
 .. _lib:
 
 ************************
-MDAnalysis math library
+MDAnalysis library
 ************************
 
-Low level access to MDAnalysis' math functions.  These have little to
+:mod:`MDAnalysis.lib._distances` contains low level access to
+MDAnalysis' math functions.  These have little to
 no error checking done on inputs so should be used with caution.
-Versions of these routines with error checking are available in 
-:module:`~MDAnalysis.core.distances`
+
+:mod:`MDAnalysis.lib.transformations` contains a multitude of
+matrix operations for manipulating coordinate data.
 
 .. toctree::
    :maxdepth: 1
 
-   ./lib/KDTree_modules
    ./lib/_distances
+   ./lib/KDTree_modules
+   ./lib/transformations
